+++ src/api/ans.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Account, AccountAddressInput } from "../core";
+import { Account, AccountAddress, AccountAddressInput } from "../core";
 import {
   RegisterNameParameters,
   getExpiration,
@@ -22,7 +22,7 @@ import {
   getDomainSubdomains,
   GetDomainSubdomainsArgs,
 } from "../internal/ans";
-import { GetANSNameResponse, MoveAddressType } from "../types";
+import { GetANSNameResponse } from "../types";
 import { InputGenerateTransactionOptions, SimpleTransaction } from "../transactions/types";
 import { AptosConfig } from "./aptosConfig";
 
@@ -43,9 +43,9 @@ export class ANS {
    *
    * @param args.name - A string of the name to retrieve
    *
-   * @returns MoveAddressType if the name is owned, undefined otherwise
+   * @returns AccountAddress if the name is owned, undefined otherwise
    */
-  async getOwnerAddress(args: { name: string }): Promise<MoveAddressType | undefined> {
+  async getOwnerAddress(args: { name: string }): Promise<AccountAddress | undefined> {
     return getOwnerAddress({ aptosConfig: this.config, ...args });
   }
 
@@ -78,9 +78,9 @@ export class ANS {
    *
    * @param args.name - A string of the name: primary, primary.apt, secondary.primary, secondary.primary.apt, etc.
    *
-   * @returns MoveAddressType if the name has a target, undefined otherwise
+   * @returns AccountAddress if the name has a target, undefined otherwise
    */
-  async getTargetAddress(args: { name: string }): Promise<MoveAddressType | undefined> {
+  async getTargetAddress(args: { name: string }): Promise<AccountAddress | undefined> {
     return getTargetAddress({ aptosConfig: this.config, ...args });
   }